requirements.txt
langchain~=0.0.349
pytube~=15.0.0
moviepy~=1.0.3
pip~=23.3.1
attrs~=23.1.0
distro~=1.8.0
wheel~=0.42.0
pillow~=10.1.0
jinja2~=3.1.2
setuptools~=68.2.2
packaging~=23.2
openai~=1.3.8
flask~=3.0.0
pinecone-client~=2.2.4
gunicorn==21.2.0
tiktoken==0.5.2